update
1008 : int -> double 수정
10430 : 수정 필요

+++ backjoon/src/backjoon/backjoon1008.java
@@ -5,8 +5,8 @@ import java.util.Scanner;
 public class backjoon1008 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int a = sc.nextInt();
-		int b = sc.nextInt();
+		double a = sc.nextDouble();
+		double b = sc.nextDouble();
 		
 		System.out.println(a/b);
